+++ src/main/java/br/com/coreeduc/adapters/configuration/RabbitMQConfiguration.java
@@ -1,0 +1,44 @@
+package br.com.coreeduc.adapters.configuration;
+
+import org.springframework.amqp.core.Binding;
+import org.springframework.amqp.core.BindingBuilder;
+import org.springframework.amqp.core.Queue;
+import org.springframework.amqp.core.TopicExchange;
+import org.springframework.amqp.support.converter.Jackson2JsonMessageConverter;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class RabbitMQConfiguration {
+    @Value("${spring.rabbitmq.queue}")
+    private String queue;
+    @Value("${rabbitmq.exchange.name}")
+    private String exchange;
+    @Value("${rabbitmq.routing.key}")
+    private String routingKey;
+
+    @Bean
+    public TopicExchange exchange(){
+        return new TopicExchange(exchange);
+    }
+
+    @Bean
+    public Binding binding(){
+        return BindingBuilder
+                .bind(queue())
+                .to(exchange())
+                .with(routingKey);
+    }
+
+    @Bean
+    public Queue queue() {
+        return new Queue(queue, true);
+    }
+
+    @Bean
+    public Jackson2JsonMessageConverter messageConverter() {
+        return new Jackson2JsonMessageConverter();
+    }
+
+}
