--- src/main/java/br/com/coreeduc/adapters/configuration/SecurityConfig.java
@@ -1,83 +1,0 @@
-package br.com.coreeduc.adapters.configuration;
-
-import br.com.coreeduc.adapters.outbound.authentication.security.JWTAuthenticationFilter;
-import br.com.coreeduc.adapters.outbound.authentication.security.JWTAuthorizationFilter;
-import br.com.coreeduc.adapters.outbound.authentication.security.JWTUtil;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.EnableAspectJAutoProxy;
-import org.springframework.http.HttpMethod;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.builders.WebSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.CorsConfigurationSource;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-
-import java.util.Arrays;
-
-@Configuration
-@EnableWebSecurity
-@EnableAspectJAutoProxy(proxyTargetClass = true)
-public class SecurityConfig extends WebSecurityConfigurerAdapter {
-
-    private static final String[] PUBLIC_MATCHERS_AUTH = {"/auth/**"};
-    private static final String[] PUBLIC_MATCHERS_AUTHORIZATION = {"/**.html", "/v2/api-docs", "/webjars/**", "/configuration/**", "/swagger-resources/**"};
-
-    @Autowired
-    private UserDetailsService userDetailsService;
-    @Autowired
-    private JWTUtil jwtUtil;
-
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-        http.addFilter(new JWTAuthenticationFilter(authenticationManager(), jwtUtil))
-                .addFilter(new JWTAuthorizationFilter(authenticationManager(), jwtUtil,  userDetailsService))
-                .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-                .and()
-                .cors()
-                .and()
-                .csrf()
-                .disable()
-                .authorizeRequests()
-                .antMatchers(HttpMethod.POST, PUBLIC_MATCHERS_AUTH)
-                .permitAll()
-                .antMatchers(HttpMethod.GET, PUBLIC_MATCHERS_AUTH)
-                .permitAll()
-                .anyRequest()
-                .authenticated();
-    }
-
-
-    @Override
-    public void configure(AuthenticationManagerBuilder auth) throws Exception {
-        auth.userDetailsService(userDetailsService).passwordEncoder(bCryptPasswordEncoder());
-    }
-
-    @Override
-    public void configure(WebSecurity web) {
-        web.ignoring().antMatchers(PUBLIC_MATCHERS_AUTHORIZATION);
-    }
-
-    @Bean
-    public CorsConfigurationSource corsConfigurationSource() {
-        var configuration = new CorsConfiguration().applyPermitDefaultValues();
-        configuration.setAllowedMethods(Arrays.asList("POST", "GET", "PUT", "DELETE", "OPTIONS"));
-        final var source = new UrlBasedCorsConfigurationSource();
-        source.registerCorsConfiguration("/**", configuration);
-        return source;
-    }
-
-    @Bean
-    public BCryptPasswordEncoder bCryptPasswordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-
-}
