+++ src/main/java/br/com/coreeduc/adapters/configuration/WebConfiguration.java
@@ -1,5 +1,6 @@
 package br.com.coreeduc.adapters.configuration;
 
+import br.com.coreeduc.adapters.outbound.configTenanty.TenantInterceptor;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
